move test back to c_array
